==> GWAS-analysis/2.snp_nexus.Rmd <==
---
title: "SNP Annotation"
output:
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = TRUE, warning = FALSE, message = TRUE)
#var <- c("gwas") 
knitr::opts_knit$set(root.dir = "~/GWAS_22/gwas_final/merge/typhoid/vast/igg")
```

#### Aims

-   Upload SNP list to SNPNexus/VEP
-   Results in various annotation files
-   Download snp annotation results as .txt files and read into R
-   For candidate gene analysis prioritise and filter SNPs with GTex tissue expression, coding snps etc.

##### Load Packages:

```{r Packages}
library(dplyr)
library(ggplot2)
library(ggrepel)
library(tidylog)
library(data.table)
library(stringr)
library(stringi)
library(openxlsx)
library(RColorBrewer)
library(forcats)
library(qqman)
library(janitor)
```

#### Write nexus snp list

Upload data to SNP nexus to get genome annotation <https://www.snp-nexus.org/v4/>

```{r format Nexus snp list, message=FALSE}

var <- c("igg_2")
tophits <- fread(paste0(var,"_tophits.txt"))
nexus <- select(tophits, CHR, BP, REF, ALT)
nexus <- mutate(nexus, strand = rep(1, length(nexus$CHR)))
nexus <- mutate(nexus, Type = rep("Chromosome", length(nexus$CHR)))
nexus <- select(nexus, Type, CHR, BP, REF, ALT, strand)
names(nexus) <-
  c("Type", "Id", "Position", "Alelle1", "Allele2", "Strand")
write.table(nexus, paste0(var,"_nexus.txt"), row.names = F, col.names = F, quote = F, sep = " ")
```

Read nexus .txt files from assoc directory into R

#### Format Annotation Data

```{r get-nexusdata, message=FALSE}
# Read in data
getwd()
temp = list.files(pattern = ".txt")
myfiles = lapply(temp, fread) #read all files in temp
trim <- str_extract(temp, ".{4,}?")
names(myfiles) <- trim

#Unlist files and save to global environment
list2env((myfiles), envir = .GlobalEnv)
tophits <- igg_
```
```{r Format Annotation data ID, message=TRUE}
# ID file -----------------------------------------------------------------------

ID <- near #contains file of nearest gene to the snps
rm(near)
ID <- ID[,1:8]

# Predicted ensembl function ----------------------------------------------------
coding_muts <-
  filter(ense,`Predicted Function` == "coding")

ense <-
  ense %>%
  dplyr::select(`Variation ID`, Variant, Symbol, `Predicted Function`) %>%
  distinct()
#Lots of duplicate entries due to contigs I think

introns <-
  ense %>% filter(str_detect(ense$`Predicted Function`, "intron"))

no_introns <-
  ense %>% filter(!str_detect(ense$`Predicted Function`, "intron"))

#collapse by ense by ID
ense <- ense %>%
  group_by(`Variation ID`, Symbol) %>%
  summarise(`Predicted Function` = toString(`Predicted Function`))

main <- full_join(ID, ense, by = "Variation ID")


# Regulatory Elements -----------------------------------------------------------
#road <- fread("roadmap_mari.johnson@lmh.ox.ac.uk.txt")
road$Epigenome <- as.factor(road$Epigenome)
#levels(road$Epigenome)
# Contains many tissue eQTLs
# Filter for immune and lung tissue expression
road <-
  road %>%
  filter(str_detect(road$Epigenome, "B cell|CD4|CD8|CD34|T cell|T helper|Monocyte|neutrophil|myeloid|Natural Killer|Spleen|macrophage|colon|Intestine")==TRUE)

# Collapse by Histone feautre
road <- road %>%
  group_by(`Variation ID`,
           `Chromosome`,
           `Feature Type Class`) %>%
  summarise(Tissue_Express = toString(Epigenome))

road$Tissue_Express <- str_remove_all(road$Tissue_Express, ("Roadmap"))
road$Tissue_Express <- str_remove_all(road$Tissue_Express, ("PB"))
road <- filter(road, Tissue_Express != "Fetal Intestine Small")
main <- left_join(main, road, by = "Variation ID")


# GWAS public data -------------------------------------------------------------

gwas <- gwas %>%
  select(`Variation ID`, Trait, `p-Value`,`Risk Allele`, PubMed)
```

```{r Join tables, message=TRUE}
# Join tables ------------------------------------------------------------------

tophits <- select(tophits, -c(TEST))

# Format SNP - VAR ID for joining
# chr1:25346917:G:A > chr1:111418204:G/A:1

tophits <- rename(tophits, `Variation ID` = SNP)
tophits$`Variation ID` <- gsub("((?:[^//:]+//:){3}[^//:]+)//:", "\\1i", tophits$`Variation ID`)

main$`Variation ID`<- str_remove_all(main$`Variation ID`, ("\\:1$"))


# Replace the third (or last) instance of :
# with // / slash
replN <- function(x,fn,rp,n) {
    sel <- rep(fn, n*length(rp))
    sel[seq_along(rp)*n] <- rp
    regmatches(x, gregexpr(fn, x)) <- list(sel)
    x
}

tophits$`Variation ID` <- replN(tophits$`Variation ID`, fn=":", rp=c("/"), n=3)
#tophits$`Variation ID` <- paste(tophits$`Variation ID`, ":1", sep = "")
tophits <- dplyr::rename(tophits, position = BP)

gwas <- clean_names(gwas)
path <- clean_names(path)
road <- clean_names(road)
ID <- clean_names(ID)
main <- clean_names(main)
main <- main %>% select(!ends_with("_y"))
names(main) <- str_replace_all(names(main),"_x","")

tophits <- clean_names(tophits)
main <- left_join(tophits, main, by = c("variation_id", "position"))


spread <- list('tophits' = main,
               'public_gwas' = gwas,
               'pathway' = path)

library(openxlsx)
write.xlsx(spread, file = "typhoid_gwas.xlsx")
write.csv(main, file = "GWAS_summary.data.csv", row.names = F)

```
#### Labelled Manhattan Plots

```{r label-data-set-up}
library(readxl)
var <- c("igg_2")
gwas_data <- clean_names(fread(paste0(var, "_gwas.txt")))
# Reduce comp time by removing dense plots at the bottom which overlap
sig_data <- gwas_data %>% 
  subset(p <= 0.01)
nsig_data <- gwas_data %>% 
  subset(p >= 0.01) %>%
  group_by(chr) %>% 
  sample_frac(0.1) # Keep 8% of nsig data
gwas_data <- bind_rows(sig_data, nsig_data)

# Calculate base pair coordinates -----------------------------
data_cum <- gwas_data %>% 
group_by(chr) %>% 
  summarise(max_bp = max(bp)) %>% 
  mutate(bp_add = lag(cumsum(as.numeric(max_bp)), default = 0)) %>% 
  select(chr, bp_add)

gwas_data <- gwas_data %>% 
  inner_join(data_cum, by = "chr") %>% 
  mutate(bp_cum = bp + bp_add)

gwas_data <- gwas_data %>%
select(-test,-bp_add)

write.csv(gwas_data, file = "igg_2_manhatt_gwas.csv")

axis_set <- gwas_data %>% 
  group_by(chr) %>% 
  summarize(center = mean(bp_cum))

ylim <- gwas_data %>% 
  filter(p == min(p)) %>% 
  mutate(ylim = abs(floor(log10(p))) + 2) %>% 
  pull(ylim)
sig <- 1e-5 # Top = 9.9e-5
```

#### Highlight genes of interest

```{r highlight genes}
# Highlight genes --------------------------------------------------------------

main <- read_excel("typhoid_gwas.xlsx", sheet = "tophits")
public <- read_excel("typhoid_gwas.xlsx", sheet = "public_gwas")

# Just remove dups for now
main <- main[(stri_duplicated(main$variation_id)==FALSE),]
# Update var_id format 
main$snp <- str_replace_all(main$variation_id, "\\/", "\\:") %>% str_remove_all("(\\:1)$")
public$snp <- str_replace_all(public$variation_id, "\\/", "\\:") %>% str_remove_all("(\\:1)$")

# Public SNPs ------------------------------------------------------------------
# Filter albumin snps for the plot
public <- public[!str_detect(public$trait, "lbumin"), ]
public$chr <- as.numeric(str_extract_all(public$snp, "^(chr\\d*)") %>% str_remove_all("chr"))
public$position <- as.numeric(str_extract_all(public$snp, "(^chr\\d*:\\d*)") %>% str_remove_all("^(chr\\d*:)"))
# get main snps for public (var id alleles were reversed)
pub_snps <- inner_join(public, main, by = c("chr", "position")) %>% select(snp.y) %>% rename(snp = snp.y)


main <- left_join(main, public, by = c("chr", "position")) %>% rename(snp = snp.x, 
                                                                      variation_id = variation_id.x )

# Top SNPs to highlight --------------------------------------------------------

# Filter so grouped by gene and then take the min val
# Just highlight top coding genes per chromosome
DT <- data.table(filter(main, type == "protein_coding"))
top_snps <- DT[, .SD[which.min(p)], by = overlapped_gene]
top_snps <- filter(top_snps, p <= 1e-4)


# EQTLs ------------------------------------------------------------------------

# eqtl <- clean_names(read.csv(file = "~/GWAS_22/gwas_final/eQTL/vast/final_eqtls/V0igaall_eqtl2.csv"))
# eqtl <- filter(eqtl, !stri_duplicated(gene))
# cis_snps <- eqtl %>% select(snp, gene, eqtl) %>% filter(eqtl == "cis")
# trans_snps <- eqtl %>% select(snp, gene, eqtl) %>% filter(eqtl == "trans")

# Get gene names and traits to highlight
gene_info <- select(main, snp, trait, overlapped_gene)
gwas_data <- left_join(gwas_data, gene_info, by = "snp")

# Add highlight and annotation information
gwas_data <- gwas_data %>%
  mutate(highlight_top = ifelse(snp %in% top_snps$snp, "yes", "no")) %>%
  mutate(highlight_pub = ifelse(snp %in% pub_snps$snp, "yes", "no")) 
 #%>% mutate(highlight_cis = ifelse(snp %in% cis_snps$snp, "yes", "no")) %>%
#  mutate(highlight_trans = ifelse(snp %in% trans_snps$snp, "yes", "no"))

```

```{r plot-manhattan public}
# Plot with labelled eqtl genes ------------------------------------------------
plot_dir <- c("~/GWAS_22/gwas_final/merge/typhoid/assoc/plots")

# table(gwas_data$highlight_top)
png(file = (paste(plot_dir, var, "-pub-genes-man.png", sep = "")), width = 10, height = 6, units = "in", res = 300)

ggplot(gwas_data,
       aes(
         x = bp_cum,
         y = -log10(p),
         color = as_factor(chr),
         size = -log10(p)
       )) +
  geom_hline(yintercept = -log10(1e-4),
             color = "grey34",
             linetype = "dashed") +
  geom_hline(yintercept = -log10(1e-8),
             color = "grey80",
             linetype = "dashed") +
  geom_point(alpha = 0.8, size = 1.3) +
  scale_x_continuous(label = axis_set$chr, breaks = axis_set$center) +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 8.5)) +
  scale_color_manual(values = rep(c("#3d8e7f", "#93d5b2"), 22)) +
  scale_size_continuous(range = c(0.5, 3)) +
  geom_label_repel(
    data = subset(gwas_data, highlight_top == "yes"),
    aes(label = overlapped_gene),
    size = 2.5, color = "#12263A", nudge_y = 1, nudge_x = 3, direction = "both"
  ) +
  geom_label_repel(
    data = subset(gwas_data, highlight_pub == "yes"),
    aes(label = trait),
    size = 2.5, color = "#C14953", nudge_y = 0.6, nudge_x = 3, direction = "both"
  ) +
    theme_minimal() +
  theme(
    legend.position = "none",
    panel.border = element_blank(),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank()
  ) +
  labs(x = "Chromsome", title = "Genetic associations of typhoid fever diagnosis")

dev.off()
save.image(file = paste0(var, "_annot.RData"))

```

```{r plot-manhattan eqtl}
# Plot with labelled eqtl genes ------------------------------------------------
plot_dir <- c("~/GWAS_22/gwas_final/merge/typhoid/vast/plots/")

png(file = (paste(plot_dir, var, "-top-genes-man.png", sep = "")), width = 10, height = 6, units = "in", res = 300)
ggplot(gwas_data,
       aes(
         x = bp_cum,
         y = -log10(p),
         color = as_factor(chr),
         size = -log10(p)
       )) +
  geom_hline(yintercept = -log10(1e-48),
             color = "grey34",
             linetype = "dashed") +
  geom_hline(yintercept = -log10(1e-8),
             color = "grey80",
             linetype = "dashed") +
  geom_point(alpha = 0.8, size = 1.3) +
  scale_x_continuous(label = axis_set$chr, breaks = axis_set$center) +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 8.5)) +
  scale_color_manual(values = rep(c("#3d8e7f", "#93d5b2"), 22)) +
  scale_size_continuous(range = c(0.5, 3)) +
  geom_label_repel(
    data = subset(gwas_data, highlight_top == "yes"),
    aes(label = overlapped_gene),
    size = 2.5, color = "#202936", nudge_y = 0.5, nudge_x = 0.1
  ) + geom_label_repel(
    data = subset(gwas_data, highlight_trans == "yes"),
    aes(label = gene),
    size = 2.5, color = "#AD3A8A",
  max.overlaps = 20, nudge_y = 0.005) +
  geom_label_repel(
    data = subset(gwas_data, highlight_cis == "yes"),
    aes(label = gene),
    size = 2.5, color = "#D28423", nudge_y = 0.1, nudge_x = 0.05
  ) +
    theme_minimal() +
  theme(
    legend.position = "none",
    panel.border = element_blank(),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank()
  ) +
  labs(x = "Chromsome", title = "Genetic associations of Vi-IgA responses following typhoid vaccination")
dev.off()

save.image(file = paste0(var, "_annot.RData"))

# Colour palette
# #fcf2c2 light colour, #e06252 light red, #202936 dark blue, #D28423 og pale yellow/orange
# +
#   geom_label_repel(
#     data = subset(gwas_data, highlight_pub == "yes"),
#     aes(label = trait),
#     size = 2.5, color = "#AD3A8A", nudge_y = -0.3, nudge_x = 2) +
#   labs(x = "Chromsome",
#        y = "-log10(p)") 

```

#### GWAS combined tophits

- check for common genes/variants
